Re-define similarity scoring for type features
Previously, to compute similarity scores of type features, we first
assigned a unique prime number to each abstract type, and then
multiplied the number of type features, and finally compared the
multiplied number between two functions using relative difference.
However, this would produce a biased result as we arbitrarily assigned a
prime number to each abstract type.

In this fix, we compute the similarity scoring of type features using
Jaccard index. We changed the name of type feature and their usage.